ring_joint_sdpa: split chunked helpers into own header to fix exp_ kernel JIT build
The chunked-prefill commit added #include "ring_utils.hpp" to
compute_common.hpp and compute_streaming.hpp so the lightweight-mask
helpers could use ChunkedContext and kv_global_tile_for_local.

That created a header diamond inside exp_ring_joint_sdpa.cpp, which
includes both compute_common.hpp / compute_streaming.hpp AND
exp_fused_op_indexer.hpp -> exp_ring_utils.hpp. The two ring_utils
headers each define their own RingIdSequencer and
find_last_active_ring_iter with different signatures, so the
transitive pull-in caused:

  error: redefinition of 'struct RingIdSequencer'
  error: no matching function for call to 'RingIdSequencer(...)'
  error: call of overloaded 'find_last_active_ring_iter(...)' is ambiguous

Move ChunkedContext + kv_global_tile_for_local into a new
chunked_prefill_utils.hpp containing nothing else. Have ring_utils.hpp
include it (so existing callers continue to see those symbols by the
same name), and switch compute_common.hpp / compute_streaming.hpp to
include only the new header — they never used RingIdSequencer or
find_last_active_ring_iter.

exp_ring_joint_sdpa.cpp now sees ChunkedContext / kv_global_tile_for_local
from chunked_prefill_utils.hpp, and RingIdSequencer / find_last_active_ring_iter
from exp_ring_utils.hpp only — no collision.

Verified: chunked accuracy sweep still passes; exp_ring_joint_sdpa
JIT-compiles cleanly (47/47 cache hits on the second exp_* run).
